Allow random state to be specified for load_candidates.
Make evaluate.candidates.modules take raw, norm so can work for cv evaluation. This required doing the same thing for data.analyse.

Add overall candidate evaluation functions for both cv and not. Both are deterministic as desired (as shuffle of test does not matter for predictions and ranks are joint).

Fixed train_predict_evaluate_cv so evaluates on combined output not each fold at a time, and added mfr benchmark. Deterministic as loading and training both passed rng instances. Also created evaluation function for not cv, which is also deterministic.